Extract `haversine` module
