+++ src/lib.rs
@@ -9,6 +9,7 @@ extern crate rayon;
 
 pub mod olc;
 pub mod flat;
+pub mod haversine;
 mod point;
 mod parallel;
