+++ src/olc.rs
@@ -4,6 +4,7 @@ use ord_subset::OrdSubsetIterExt;
 
 use crate::Point;
 use crate::flat::to_flat_points;
+use crate::haversine::haversine_distance;
 use crate::parallel::*;
 
 const LEGS: usize = 6;
@@ -125,21 +126,4 @@ fn calculate_distance<T: Point>(route: &[T], point_list: &[usize]) -> f64 {
         .map(|(i1, i2)| (&route[i1], &route[i2]))
         .map(|(fix1, fix2)| haversine_distance(fix1, fix2))
         .sum()
-}
-
-fn haversine_distance(fix1: &Point, fix2: &Point) -> f64 {
-    const R: f64 = 6371.; // kilometres
-
-    let phi1 = fix1.latitude().to_radians();
-    let phi2 = fix2.latitude().to_radians();
-    let delta_phi = (fix2.latitude() - fix1.latitude()).to_radians();
-    let delta_rho = (fix2.longitude() - fix1.longitude()).to_radians();
-
-    let a = (delta_phi / 2.).sin() * (delta_phi / 2.).sin() +
-        phi1.cos() * phi2.cos() *
-            (delta_rho / 2.).sin() * (delta_rho / 2.).sin();
-
-    let c = 2. * a.sqrt().atan2((1. - a).sqrt());
-
-    R * c
 }
